メモリ節約とGPSの値保持

+++ Arduino/Test/Camera.ino
@@ -58,7 +58,7 @@ void CAM_TakePic() {
 	// Create an image with the name IMGxx.JPG
 	char filename[12];
 	strcpy(filename, "000.JPG");
-	for (int i = 0; i < 1000; i++) {
+	for (uint16_t i = 0; i < 1000; i++) {
 		filename[0] = '0' + i/100;
 		filename[1] = '0' + (i/10)%10;
 		filename[2] = '0' + i%10;
